requirements.txt
Flask==3.0.3
requests==2.32.3
gspread==6.1.2
google-auth==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
